Move feature selection functionality to separate file
